Added removing content type draft when clicking cancel button
