+++ src/bundle/Resources/views/themes/admin/content_type/create.html.twig
@@ -6,8 +6,7 @@
 
 {% block header %}
     {% set content_type_create_sidebar_right = knp_menu_get('ezplatform_admin_ui.menu.content_type_create.sidebar_right', [], {
-        'save_id': form.publishContentType.vars.id,
-        'group': content_type_group
+        'form_view': form,
     }) %}
 
     {% include '@ibexadesign/ui/edit_header.html.twig' with {
@@ -86,6 +85,7 @@
         </div>
         {{ form_widget(form.saveContentType, { attr: { hidden: 'hidden' }}) }}
         {{ form_widget(form.publishContentType, { attr: { class: 'ibexa-content-type-edit__publish-content-type', hidden: 'hidden' }}) }}
+        {{ form_widget(form.removeDraft, { attr: { hidden: 'hidden', formnovalidate: true }}) }}
 
         {{ form_widget(form._token) }}
     {{ form_end(form, {'render_rest': false }) }}
