+++ src/lib/Menu/Admin/ContentType/ContentTypeCreateRightSidebarBuilder.php
@@ -7,88 +7,22 @@
 namespace Ibexa\AdminUi\Menu\Admin\ContentType;
 
 use Ibexa\AdminUi\Menu\Event\ConfigureMenuEvent;
-use Ibexa\AdminUi\Menu\MenuItemFactory;
-use Ibexa\Contracts\AdminUi\Menu\AbstractBuilder;
-use Ibexa\Contracts\Core\Repository\Exceptions as ApiExceptions;
 use JMS\TranslationBundle\Model\Message;
-use JMS\TranslationBundle\Translation\TranslationContainerInterface;
-use Knp\Menu\ItemInterface;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * KnpMenuBundle Menu Builder service implementation for AdminUI Section Edit contextual sidebar menu.
  *
  * @see https://symfony.com/doc/current/bundles/KnpMenuBundle/menu_builder_service.html
  */
-class ContentTypeCreateRightSidebarBuilder extends AbstractBuilder implements TranslationContainerInterface
+class ContentTypeCreateRightSidebarBuilder extends AbstractContentTypeRightSidebarBuilder
 {
     /* Menu items */
     public const ITEM__SAVE = 'content_type_create__sidebar_right__save';
     public const ITEM__CANCEL = 'content_type_create__sidebar_right__cancel';
 
-    /** @var \Symfony\Contracts\Translation\TranslatorInterface */
-    private $translator;
-
-    public function __construct(
-        MenuItemFactory $factory,
-        EventDispatcherInterface $eventDispatcher,
-        TranslatorInterface $translator
-    ) {
-        parent::__construct($factory, $eventDispatcher);
-
-        $this->translator = $translator;
-    }
-
-    /**
-     * @return string
-     */
     protected function getConfigureEventName(): string
     {
         return ConfigureMenuEvent::CONTENT_TYPE_CREATE_SIDEBAR_RIGHT;
-    }
-
-    /**
-     * @param array $options
-     *
-     * @return \Knp\Menu\ItemInterface
-     *
-     * @throws \InvalidArgumentException
-     * @throws ApiExceptions\BadStateException
-     * @throws \InvalidArgumentException
-     */
-    public function createStructure(array $options): ItemInterface
-    {
-        /** @var \Ibexa\Contracts\Core\Repository\Values\ContentType\ContentTypeGroup $section */
-        $group = $options['group'];
-
-        $saveId = $options['save_id'];
-
-        /** @var \Knp\Menu\ItemInterface|\Knp\Menu\ItemInterface[] $menu */
-        $menu = $this->factory->createItem('root');
-
-        $menu->setChildren([
-            self::ITEM__SAVE => $this->createMenuItem(
-                self::ITEM__SAVE,
-                [
-                    'attributes' => [
-                        'class' => 'ibexa-btn--trigger ibexa-btn--save-content-type',
-                        'data-click' => sprintf('#%s', $saveId),
-                    ],
-                ]
-            ),
-            self::ITEM__CANCEL => $this->createMenuItem(
-                self::ITEM__CANCEL,
-                [
-                    'route' => 'ibexa.content_type_group.view',
-                    'routeParameters' => [
-                        'contentTypeGroupId' => $group->id,
-                    ],
-                ]
-            ),
-        ]);
-
-        return $menu;
     }
 
     /**
@@ -100,6 +34,16 @@ class ContentTypeCreateRightSidebarBuilder extends AbstractBuilder implements Tr
             (new Message(self::ITEM__SAVE, 'menu'))->setDesc('Create'),
             (new Message(self::ITEM__CANCEL, 'menu'))->setDesc('Cancel'),
         ];
+    }
+
+    public function getItemSaveIdentifier(): string
+    {
+        return self::ITEM__SAVE;
+    }
+
+    public function getItemCancelIdentifier(): string
+    {
+        return self::ITEM__CANCEL;
     }
 }
